Code midified to use COnditional varibale. Transfer thread only runs if condition set from Poller thread. Poller and Logger thread synchronise using semaphore.

+++ Common/Common.h
@@ -20,14 +20,14 @@
 #include <string.h>
 
 //******************************* Global Types ********************************
-typedef signed char             int8;
+typedef char             int8;
 typedef unsigned char           uint8;
 typedef short                   int16;
 typedef unsigned short          uint16;
 typedef int                     int32;
-typedef unsigned long           uint32;
-typedef long long               int64;
-typedef unsigned long long      uint64;
+typedef unsigned int            uint32;
+typedef long int                int64;
+typedef unsigned long int       uint64;
 
 //***************************** Global Constants ******************************
 #define RETURN_OK               0
